Update build scripts
