+++ .github/workflows/build.yaml
@@ -13,10 +13,12 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v2
 
-      - name: Set up JDK 13
-        uses: actions/setup-java@v1
+      - name: Set up JDK 18
+        uses: actions/setup-java@v3
         with:
-          java-version: 13.0.1
+          distribution: corretto
+          java-version: 18
+          cache: gradle
 
       - name: Gradle Build
         run: |
