+++ .github/workflows/release.yaml
@@ -11,18 +11,20 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
         with:
           fetch-depth: 0
 
       - name: Calculate version
         id: spell
-        uses: actoaps/version-spell@v2
+        uses: actoaps/version-spell@v3
 
-      - name: Set up JDK 13
-        uses: actions/setup-java@v1
+      - name: Set up JDK 18
+        uses: actions/setup-java@v3
         with:
-          java-version: 13.0.1
+          distribution: corretto
+          java-version: 18
+          cache: gradle
 
       - name: Gradle Build
         run: |
@@ -32,14 +34,25 @@ jobs:
           SONAR_LOGIN: ${{ secrets.SONAR_LOGIN }}
           BUILD_ID: ${{ steps.spell.outputs.version }}
 
-      - name: Release to Docker
-        run: |
-          docker login --username $DOCKER_HUB_USERNAME --password $DOCKER_HUB_PASSWORD
-          docker build -t actoaps/chromium-pdf-gen:$BUILD_ID -t actoaps/chromium-pdf-gen:latest .
+      - name: Set up QEMU for multiarch builds
+        uses: docker/setup-qemu-action@v2
+        with:
+          platforms: arm64/v8,amd64
 
-          docker push actoaps/chromium-pdf-gen:$BUILD_ID
-          docker push actoaps/chromium-pdf-gen:latest
-        env:
-          DOCKER_HUB_USERNAME: ${{ secrets.DOCKER_HUB_USERNAME }}
-          DOCKER_HUB_PASSWORD: ${{ secrets.DOCKER_HUB_PASSWORD }}
-          BUILD_ID: ${{ steps.spell.outputs.version }}
+      - name: Set up Docker Buildx
+        id: buildx
+        uses: docker/setup-buildx-action@v2
+
+      - name: Login to DockerHub
+        uses: docker/login-action@v2
+        with:
+          username: ${{ secrets.DOCKER_HUB_USERNAME }}
+          password: ${{ secrets.DOCKER_HUB_PASSWORD }}
+
+      - name: Build sso image and publish to Docker
+        uses: docker/build-push-action@v3
+        with:
+          context: .
+          platforms: linux/amd64,linux/arm64/v8
+          push: true
+          tags: actoaps/chromium-pdf-gen:${{ steps.spell.outputs.version }},actoaps/chromium-pdf-gen:latest
